fix jackett

+++ scripts/update/jackett.sh
@@ -3,9 +3,10 @@
 #
 # SERVIDOR HD
 
+
+if [[ -f /install/.jackett.lock ]]; then
   username=$(cut -d: -f1 < /root/.master.info)
   active=$(systemctl is-active jackett@$username)
-
   if grep -q "ExecStart=/usr/bin/mono" /etc/systemd/system/jackett@.service; then
     jackettver=$(wget -q https://github.com/Jackett/Jackett/releases/latest -O - | grep -E \/tag\/ | grep -v repository | awk -F "[><]" '{print $3}')
     sed -i 's|ExecStart.*|ExecStart=/bin/sh -c "/home/%I/Jackett/jackett_launcher.sh"|g' /etc/systemd/system/jackett@.service
@@ -62,18 +63,18 @@ while pgrep -u ${user} JackettUpdater > /dev/null ; do
 done
 echo "Jackett update complete"
 JL
+
     chmod +x /home/${username}/Jackett/jackett_launcher.sh
 
     if [[ $active == "active" ]]; then
       restartjackett=1
     fi
   fi
-  
+
   if grep -q "proxy_set_header" /etc/nginx/apps/jackett.conf; then
     sed -i "/proxy_set_header/d" /etc/nginx/apps/jackett.conf
     systemctl reload nginx
   fi
-
   if [[ $restartjackett == 1 ]]; then
     systemctl restart jackett@${username}
   fi
